requirements.txt
fastapi[all]==0.88.0
uvicorn[standard]==0.20.0
Pillow==9.3.0
requests==2.28.1
boto3==1.26.37
python-decouple==3.6
python-socketio[asyncio_client]==5.7.2
python-socketio==5.7.2